requirements.txt
gspread==5.12.0
google-auth==2.33.0
